reorganize files into respective folders
