+++ src/index.ts
@@ -34,7 +34,7 @@ export type {TouchOrMouseEvent} from './hooks/useOnOutsideClick'
 export {useOpenAndCloseFocus} from './hooks/useOpenAndCloseFocus'
 export {useOnEscapePress} from './hooks/useOnEscapePress'
 export {useOverlay} from './hooks/useOverlay'
-export {useConfirm} from './Dialog/ConfirmationDialog'
+export {useConfirm} from './ConfirmationDialog/ConfirmationDialog'
 export {useFocusTrap} from './hooks/useFocusTrap'
 export type {FocusTrapHookSettings} from './hooks/useFocusTrap'
 export {useFocusZone} from './hooks/useFocusZone'
@@ -88,8 +88,8 @@ export {default as Details} from './Details'
 export type {DetailsProps} from './Details'
 export {default as Dialog} from './Dialog'
 export type {DialogProps, DialogHeaderProps} from './Dialog'
-export type {ConfirmationDialogProps} from './Dialog/ConfirmationDialog'
-export {ConfirmationDialog} from './Dialog/ConfirmationDialog'
+export type {ConfirmationDialogProps} from './ConfirmationDialog/ConfirmationDialog'
+export {ConfirmationDialog} from './ConfirmationDialog/ConfirmationDialog'
 export {default as FilteredSearch} from './FilteredSearch'
 export type {FilteredSearchProps} from './FilteredSearch'
 export {default as FilterList} from './FilterList'
